Shop Controller Tests

+++ src/main/java/com/amr/project/service/impl/ShopServiceImpl.java
@@ -2,6 +2,8 @@ package com.amr.project.service.impl;
 
 import com.amr.project.converter.mappers.ShopMapper;
 import com.amr.project.dao.ShopRepository;
+import com.amr.project.exception.ShopAlreadyExistingException;
+import com.amr.project.exception.ShopNotFoundException;
 import com.amr.project.model.dto.ShopDto;
 import com.amr.project.model.entity.Shop;
 import com.amr.project.service.abstracts.ShopService;
@@ -40,20 +42,29 @@ public class ShopServiceImpl implements ShopService {
     }
 
     @Override
-    public void updateShopById(ShopDto shop) {
-        Shop shop1 = shopMapper.toEntity(shop, new CycleAvoidingMappingContext());
-        shopRepository.saveAndFlush(shop1);
+    public void updateShopById(ShopDto shopDto) {
+        Shop shop = shopMapper.toEntity(shopDto, new CycleAvoidingMappingContext());
+        if (!shopRepository.existsById(shopDto.getId())) {
+            throw new ShopNotFoundException();
+        }
+        shopRepository.saveAndFlush(shop);
     }
 
     @Override
     public void deleteShopById(Long id) {
+        if (!shopRepository.existsById(id)) {
+            throw new ShopNotFoundException();
+        }
         shopRepository.deleteById(id);
     }
 
     @Override
-    public void saveShop(ShopDto shop) {
-        Shop shop1 = shopMapper.toEntity(shop, new CycleAvoidingMappingContext());
-        shopRepository.saveAndFlush(shop1);
+    public void saveShop(ShopDto shopDto) {
+        Shop shop = shopMapper.toEntity(shopDto, new CycleAvoidingMappingContext());
+        if (shopRepository.existsByName(shop)) {
+            throw new ShopAlreadyExistingException("Shop with Name " + shop.getName() + " already exists");
+        }
+        shopRepository.saveAndFlush(shop);
     }
 
     @Override
